move type def, del needless import/const
