+++ server/src/chatItem.ts
@@ -1,20 +1,3 @@
-export type Room = {
-  id: string;
-  title: string;
-  topics: Topic[];
-};
-
-export type Topic = {
-  id: string;
-  title: string;
-  description: string;
-  urls: Partial<Record<TopicLinkType, string | undefined>>;
-};
-
-export type TopicLinkType = "github" | "slide" | "product";
-
-export type TopicState = "not-started" | "active" | "paused" | "finished";
-
 export type ChatItemBase = {
   id: string;
   topicId: string;
