+++ server/src/ioServer.ts
@@ -16,7 +16,6 @@ const createSocketIOServer = (httpServer: HttpServer) => {
 
   const rooms: Record<string, RoomClass> = {};
   let activeUserCount: number = 0;
-
   let serverAwakerTimer: NodeJS.Timeout;
 
   //サーバー起こしておくため
@@ -52,7 +51,11 @@ const createSocketIOServer = (httpServer: HttpServer) => {
       socket.on("ADMIN_BUILD_ROOM", (received, callback) => {
         try {
           const roomId = uuid();
-          const newRoom = new RoomClass(roomId, received.title, received.topics);
+          const newRoom = new RoomClass(
+            roomId,
+            received.title,
+            received.topics
+          );
           rooms[roomId] = newRoom;
           console.log(`new room build: ${roomId}`);
           callback({
@@ -116,7 +119,10 @@ const createSocketIOServer = (httpServer: HttpServer) => {
             activeUserCount: room.activeUserCount,
           });
         } catch (e) {
-          console.log(`${e.message ?? "Unknown error."} (ENTER_ROOM)`, new Date().toISOString());
+          console.log(
+            `${e.message ?? "Unknown error."} (ENTER_ROOM)`,
+            new Date().toISOString()
+          );
         }
       });
 
@@ -124,7 +130,9 @@ const createSocketIOServer = (httpServer: HttpServer) => {
       socket.on("ADMIN_START_ROOM", (_) => {
         try {
           if (roomId == null) {
-            throw new Error("[sushi-chat-server] You do not joined in any room");
+            throw new Error(
+              "[sushi-chat-server] You do not joined in any room"
+            );
           }
           const room = rooms[roomId];
           room.startRoom();
@@ -140,7 +148,9 @@ const createSocketIOServer = (httpServer: HttpServer) => {
       socket.on("ADMIN_CHANGE_TOPIC_STATE", (received) => {
         try {
           if (roomId == null) {
-            throw new Error("[sushi-chat-server] You do not joined in any room");
+            throw new Error(
+              "[sushi-chat-server] You do not joined in any room"
+            );
           }
           const room = rooms[roomId];
           room.changeTopicState(received);
@@ -167,7 +177,9 @@ const createSocketIOServer = (httpServer: HttpServer) => {
           );
 
           if (roomId == null) {
-            throw new Error("[sushi-chat-server] You do not joined in any room");
+            throw new Error(
+              "[sushi-chat-server] You do not joined in any room"
+            );
           }
           const room = rooms[roomId];
 
@@ -184,12 +196,17 @@ const createSocketIOServer = (httpServer: HttpServer) => {
       socket.on("POST_STAMP", (params) => {
         try {
           if (roomId == null) {
-            throw new Error("[sushi-chat-server] You do not joined in any room");
+            throw new Error(
+              "[sushi-chat-server] You do not joined in any room"
+            );
           }
           const room = rooms[roomId];
           room.postStamp(socket.id, params);
         } catch (e) {
-          console.log(`${e.message ?? "Unknown error."} (POST_STAMP)`, new Date().toISOString());
+          console.log(
+            `${e.message ?? "Unknown error."} (POST_STAMP)`,
+            new Date().toISOString()
+          );
         }
       });
 
@@ -197,7 +214,9 @@ const createSocketIOServer = (httpServer: HttpServer) => {
       socket.on("ADMIN_FINISH_ROOM", () => {
         try {
           if (roomId == null) {
-            throw new Error("[sushi-chat-server] You do not joined in any room");
+            throw new Error(
+              "[sushi-chat-server] You do not joined in any room"
+            );
           }
           const room = rooms[roomId];
           room.finishRoom();
@@ -213,7 +232,9 @@ const createSocketIOServer = (httpServer: HttpServer) => {
       socket.on("ADMIN_CLOSE_ROOM", () => {
         try {
           if (roomId == null) {
-            throw new Error("[sushi-chat-server] You do not joined in any room");
+            throw new Error(
+              "[sushi-chat-server] You do not joined in any room"
+            );
           }
           const room = rooms[roomId];
           room.closeRoom();
