+++ server/src/topic.ts
@@ -1,5 +1,10 @@
 export type Topic = {
-  id: string; // トピックID
-  title: string; // トピックのタイトル
-  roomid: any;
+  id: string;
+  title: string;
+  description: string;
+  urls: Partial<Record<TopicLinkType, string | undefined>>;
 };
+
+export type TopicLinkType = "github" | "slide" | "product";
+
+export type TopicState = "not-started" | "active" | "paused" | "finished";
